// web/types.d.ts
export type RenderText = {
    format?: string;
    text?: string;
    env?: string;
    font: string;
    color: string;
    when?: "hover" | "click" | undefined;
};
export type WidgetStateMeta = {
    state: string;
    overlay: string;
    bbox: [number, number, number, number];
    render_text?: RenderText;
    region?: {
        mode: "cells" | "full";
        cells?: number[][];
    };
};
export type WidgetMeta = {
    id: string;
    states: WidgetStateMeta[];
};